Added xlsx storage and export definitions.
Including XLConnect is a source of considerable pain.  Roxygen requires R>=2.14
(we have 2.15) and XLConnect is not yet available for 2.15. That should not
pose any problem for production systems but is annoying on the development
side.

For now, the library call to XLConnect is not listed as a dependency as this
would invalidate all checks; we must live with the warning for now.

Signed-off-by: Marc Kirchner <mail@marc-kirchner.de>